# BitOperations arm64 intrinsic for LeadingZeroCount, TrailingZeroCount and Log2