@BruceForstall 2020-04-03
The `Log2_uint` improvement % number looks incorrect.

For these arm64-specific libraries changes, I think you need to trigger the runtime AzDO pipeline manually, as the default PR testing does not include libraries arm64 testing.