@EgorBo 2020-04-03
What about `BitOperations.RotateLeft` does jit lower GT_ROR to arm's `ror` ?