@kunalspathak 2020-04-03
> What about `BitOperations.RotateLeft` does jit lower GT_ROR to arm's `ror` ?

Yes. It converts `RotateRight` to `ror` and `RotateLeft` to `neg/ror` because there is no `rol` instruction in arm.