@BruceForstall 2020-04-04
> @BruceForstall - This measurements were taken on ARM64 machine manually for MicroBenchmarks.

given this:

API | before (in us) | after (in us) | % diff
-- | -- | -- | --
Log2_uint | 5.774 | 1.920 | 15%
Log2_ulong | 6.925 | 1.921 | 72%

It seems like `Log2_unit` should be 67% improvement, not 15%.