@kunalspathak 2020-04-06
> It seems like `Log2_unit` should be 67% improvement, not 15%.

Ah yes.